Updated dockerfile + docker compose to fix dependencies
Right each container, for example gateway has the complete system/
folder. it does an innate go download to download all the dependencies
as stated by the go.mod and go.sum. We should update this in the future
to only copy the neccesary files (The best case would be to have a base
image that initializes the go.mod, does the required go.get and then in
the specific gofile copy the worker.